+++ test/test_bivy_indexable.rb
@@ -4,20 +4,7 @@ require "test_helper"
 require "active_record"
 
 class TestBivyIndexable < Minitest::Test
-  def setup
-    # Clear any previously tracked indexable models
-    Bivy::Indexable.instance_variable_set(:@models, Set.new) if Bivy::Indexable.instance_variable_defined?(:@models)
-
-    # Define test models for each test to avoid conflicts
-    @test_models = []
-  end
-
-  def teardown
-    # Clean up any test models created during tests
-    @test_models.each do |model_name|
-      Object.send(:remove_const, model_name) if Object.const_defined?(model_name)
-    end
-  end
+  include IndexableTestHelper
 
   def test_bivy_indexable_concern_exists
     assert defined?(Bivy::Indexable), "Bivy::Indexable concern should be defined"
@@ -101,21 +88,5 @@ class TestBivyIndexable < Minitest::Test
 
     assert first_call.include?(model_class), "First call should include the model"
     assert second_call.include?(model_class), "Second call should still include the model"
-  end
-
-  private
-
-  def create_test_model(name, &block)
-    # Create a new ActiveRecord model class for testing
-    model_class = Class.new(ActiveRecord::Base) do
-      self.table_name = "campgrounds"
-      class_eval(&block) if block_given?
-    end
-
-    # Assign it to a constant so it has a name
-    Object.const_set(name, model_class)
-    @test_models << name
-
-    model_class
   end
 end